social_media_api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
django-filter==23.3
psycopg2-binary==2.9.7
dj-database-url==2.1.0
boto3==1.28.57
django-storages==1.13.2
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
django-on-heroku==1.1.2